[#94] Feat: 헤더 mini menu component 추가
